requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
Django==5.0.1
Flask==3.0.1
Flask-Login==0.6.3
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
whitenoise==6.6.0
xlwt==1.3.0
python-dotenv==1.0.1
pymysql==1.1.0
